Change project structure. Refactor points in customer table

+++ src/main/resources/db/migration/V2__AddCustomerTable.sql
@@ -5,8 +5,8 @@ create table customers
     id          bigserial primary key,
     telegramId  bigint not null,
     yearOfBirth date,
-    gender      genderType     default 'UNKNOWN',
-    points      decimal(12, 3) default 0,
+    gender      genderType default 'UNKNOWN',
+    points      int        default 0,
     updatedAt   timestamp,
     createdAt   timestamp
 );
